Create project_practicing_loops_linear_regression.py
This project is implementing a version of Linear Regression in Python. The program is running an experiment to bounce different sizes of balls, and fitting lines to the data points. This algorithm minimizes the error (the distance from each point to the line), and finds the best fit when given a set of data.